Change Data type of price from Int to Decimal

+++ Source code/backend/Repositories/Model/Service.cs
@@ -7,7 +7,7 @@ public class Service {
     public string ServiceID { get; set; }
     public string ServiceDeliveryMethodID { get; set; }
     public string Name { get; set; }
-    public int Price { get; set; }
+    public Decimal Price { get; set; }
     public string Description { get; set; }
     public string Status { get; set; }
